Tests/LibWeb/Text/expected/HTML/HTMLVideoElement-resize-event-during-playback.txt

resize: videoWidth=320 videoHeight=240
loadedmetadata: videoWidth=320 videoHeight=240
--- play ---
resize: videoWidth=640 videoHeight=480
ended: videoWidth=640 videoHeight=480
--- seek to 0 ---
resize: videoWidth=320 videoHeight=240
seeked: videoWidth=320 videoHeight=240
--- seek to end ---
resize: videoWidth=640 videoHeight=480
ended: videoWidth=640 videoHeight=480
seeked: videoWidth=640 videoHeight=480
